requirements.txt
Flask==3.0.3
google_api_python_client==2.142.0
oauth2client==4.1.3
python-dotenv==1.0.1
